+++ locales/ruRU.lua
@@ -9,4 +9,4 @@ local L = LibStub("AceLocale-3.0"):NewLocale("MyLootHistory", "ruRU")
 if not L then return end
 
 -- core
-L["_MoneyPattern"] = "%d+ "
+L["_MoneyPattern"] = "%d+"